Refactor TestClassModelBuilder into focused files (#9910)
Split the 933-line TestClassModelBuilder.cs into smaller, single-responsibility files within the same Generators namespace:

- DynamicDataSourceBuilder: [DynamicData] source resolution
- AttributeMaterializationHelper: attribute collection/materialization + TypedConstant conversion
- DataRowBuilder: [DataRow] parsing
- SymbolReferenceabilityHelper: reusable accessibility/referenceability predicates
- TestMemberValidationHelper: member eligibility, unsupported-shape reporting, signature keys
- SymbolDisplayFormats: shared fully-qualified SymbolDisplayFormat

TestClassModelBuilder now only holds top-level orchestration. Behavior is unchanged; all files are under 300 lines. Updated the BuildAttributes call site in MSTestReflectionMetadataGenerator.

Copilot-Session: 2fc1a0b3-faab-4c48-a40a-23eae6fa9386
